djangoapi/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.4
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
requests==2.32.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
six==1.17.0
debugpy==1.8.11
cryptography==44.0.0
django-rest-knox==5.0.2
pgOperations==0.0.3
Datetime==5.5
django-filter==24.3
drf-access-policy==1.5.0
drf-yasg
djangorestframework-gis